@@ -4,8 +4,8 @@ FactoryGirl.define do
     location_name 'Westminster'
     address { Faker::Address::street_address }
     event_date { Date.today }
-    start_time { Faker::Time.between(2.days.ago, Time.now, :afternoon) }
-    end_time { Faker::Time.between(2.days.ago, Time.now, :evening) }
+    start_time { Faker::Time.between(2.days.ago, Time.zone.now, :afternoon) }
+    end_time { Faker::Time.between(2.days.ago, Time.zone.now, :evening) }
     description { Faker::Lorem.sentence }
     location_url { Faker::Internet.url }
   end
